Extend Decidim Comments Comment model
