+++ lib/decidim/transparent_trash.rb
@@ -10,6 +10,7 @@ require "decidim/transparent_trash/extends/initiatives_permissions"
 require "decidim/transparent_trash/extends/initiatives_admin_permissions"
 require "decidim/transparent_trash/extends/initiatives_controller"
 require "decidim/transparent_trash/extends/initiatives_admin_controller"
+require "decidim/transparent_trash/extends/comments_seed"
 
 module Decidim
   # This namespace holds the logic of the `TransparentTrash` component. This component
