+++ lib/decidim/transparent_trash/engine.rb
@@ -20,11 +20,16 @@ module Decidim
       end
 
       initializer "decidim.transparent_trash.extends" do
+        # Commands
         Decidim::Initiatives::Admin::UnpublishInitiative.include Decidim::TransparentTrash::Extends::UnpublishInitiative
+        # Permissions
         Decidim::Initiatives::Permissions.include Decidim::TransparentTrash::Extends::InitiativesPermissions
         Decidim::Initiatives::Admin::Permissions.include Decidim::TransparentTrash::Extends::InitiativesAdminPermissions
+        # Controllers
         Decidim::Initiatives::InitiativesController.include Decidim::TransparentTrash::Extends::InitiativesController
         Decidim::Initiatives::Admin::InitiativesController.include Decidim::TransparentTrash::Extends::InitiativesAdminController
+        # Models
+        Decidim::Comments::Seed.include Decidim::TransparentTrash::Extends::CommentsSeed
       end
     end
   end
